[#10] refactor: page복구/오타수정등
