+++ src/app/page.tsx
@@ -1,12 +1,1 @@
-import BookingInterface from '@/components/floating-box/BookingInterface';
-
-export default function Home() {
-  return (
-    <>
-      <BookingInterface />
-      <BookingInterface />
-      <BookingInterface />
-      <BookingInterface />
-    </>
-  );
-}
+export default function Home() {}
