+++ src/hooks/useDeviceSize.ts
@@ -2,8 +2,15 @@ import { useEffect, useState } from 'react';
 
 type DeviceType = 'mobile' | 'tablet' | 'desktop';
 
-export default function useDeviceType(): DeviceType {
-  const [device, setDevice] = useState<DeviceType>('mobile');
+export default function useDeviceSize(): DeviceType {
+  const getDeviceType = () => {
+    const width = window.innerWidth;
+    if (width < 768) return 'mobile';
+    else if (width < 1024) return 'tablet';
+    else return 'desktop';
+  };
+
+  const [device, setDevice] = useState<DeviceType>(getDeviceType());
 
   useEffect(() => {
     const checkDevice = () => {
