+++ src/ui/BookingModal.tsx
@@ -3,9 +3,10 @@
 import useDeviceType from '@/hooks/useDeviceSize';
 import MobileModal from './MobileBookingModal';
 import TabletModal from './TabletBookingModal';
+import useDeviceSize from '@/hooks/useDeviceSize';
 
 export default function BookingModal() {
-  const device = useDeviceType();
+  const device = useDeviceSize();
 
   if (device === 'mobile') return <MobileModal />;
   if (device === 'tablet') return <TabletModal />;
